Better strategy for tab selection on close.
Before, TabStrip was tracking tab selection history and
backtracking when a tab was closed. This was stupid and led
to O(N^2) tab selection when closing the current tab.

Now views are simply updated with a timestamp any time they
are activated, and when the current tab is closed, the most
recently used view is activated in its place. Much better.